# 前端/myoj/openapitools.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "通用后端模版 API 文档",
    "description": "登录、注册、用户管理等接口",
    "contact": {
      "name": "开发团队",
      "email": "dev@example.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888/api",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/question_submit/": {
      "post": {
        "tags": [
          "question-submit-controller"
        ],
        "operationId": "doQuestionSubmit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionSubmitAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultQuestionSubmitCreatedVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/question/admin/page": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "管理员分页查询题目",
        "description": "支持按标题、内容、标签、出题人筛选",
        "operationId": "adminPageQuestions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageQuestionAdminListItemVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/question_submit/status/{id}": {
      "get": {
        "tags": [
          "question-submit-controller"
        ],
        "operationId": "getSubmitStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultMySubmitItemVO"
                }
              }
            }
          }
        }
      }
    },
    "/question_submit/status/my": {
      "get": {
        "tags": [
          "question-submit-controller"
        ],
        "operationId": "getSubmitStatusBySubmitNo",
        "parameters": [
          {
            "name": "questionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "submitNo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultMySubmitItemVO"
                }
              }
            }
          }
        }
      }
    },
    "/question_submit/my/page/all": {
      "get": {
        "tags": [
          "question-submit-controller"
        ],
        "operationId": "pageMyAllQuestionSubmits",
        "parameters": [
          {
            "name": "submitNo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "current",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageMySubmitItemVO"
                }
              }
            }
          }
        }
      }
    },
    "/question_submit/admin/page": {
      "get": {
        "tags": [
          "question-submit-controller"
        ],
        "operationId": "pageAllQuestionSubmitsForAdmin",
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "questionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "submitNo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageMySubmitItemVO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "QuestionSubmitAddRequest": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "questionId": {
            "type": "integer",
            "format": "int64"
          },
          "contestId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QuestionSubmitCreatedVO": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "format": "int64"
          },
          "submitNo": {
            "type": "integer",
            "format": "int64"
          },
          "globalSubmitNo": {
            "type": "integer",
            "format": "int64"
          },
          "dbId": {
            "type": "string"
          }
        }
      },
      "ResultQuestionSubmitCreatedVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/QuestionSubmitCreatedVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "QuestionQueryRequest": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          },
          "sortField": {
            "type": "string"
          },
          "sortOrder": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "content": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userNickname": {
            "type": "string"
          }
        }
      },
      "JudgeCase": {
        "type": "object",
        "properties": {
          "input": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          },
          "output": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          }
        },
        "required": [
          "input",
          "output"
        ]
      },
      "JudgeConfig": {
        "type": "object",
        "properties": {
          "timeLimit": {
            "type": "integer",
            "format": "int64",
            "maximum": 10000
          },
          "memoryLimit": {
            "type": "integer",
            "format": "int64",
            "maximum": 1048576
          },
          "stackLimit": {
            "type": "integer",
            "format": "int64",
            "maximum": 262144
          }
        },
        "required": [
          "memoryLimit",
          "stackLimit",
          "timeLimit"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "column": {
            "type": "string"
          },
          "asc": {
            "type": "boolean"
          }
        }
      },
      "PageQuestionAdminListItemVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAdminListItemVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "writeOnly": true
          },
          "optimizeCountSql": {
            "writeOnly": true
          },
          "searchCount": {
            "writeOnly": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "writeOnly": true
          },
          "maxLimit": {
            "type": "integer",
            "format": "int64",
            "writeOnly": true
          },
          "countId": {
            "type": "string",
            "writeOnly": true
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        }
      },
      "QuestionAdminListItemVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "answer": {
            "type": "string"
          },
          "submitNum": {
            "type": "integer",
            "format": "int32"
          },
          "acceptedNum": {
            "type": "integer",
            "format": "int32"
          },
          "judgeCase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JudgeCase"
            }
          },
          "judgeConfig": {
            "$ref": "#/components/schemas/JudgeConfig"
          },
          "thumbNum": {
            "type": "integer",
            "format": "int32"
          },
          "favourNum": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userNickname": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultPageQuestionAdminListItemVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageQuestionAdminListItemVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "MySubmitItemVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "submitNo": {
            "type": "integer",
            "format": "int64"
          },
          "globalSubmitNo": {
            "type": "integer",
            "format": "int64"
          },
          "language": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "contestId": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "judgeInfo": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultMySubmitItemVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/MySubmitItemVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PageMySubmitItemVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MySubmitItemVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "writeOnly": true
          },
          "optimizeCountSql": {
            "writeOnly": true
          },
          "searchCount": {
            "writeOnly": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "writeOnly": true
          },
          "maxLimit": {
            "type": "integer",
            "format": "int64",
            "writeOnly": true
          },
          "countId": {
            "type": "string",
            "writeOnly": true
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        }
      },
      "ResultPageMySubmitItemVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageMySubmitItemVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}